deployment/scripts/requirements.txt
google-cloud-secret-manager>=2.0.0
click>=8.0.0
